final/src/requirement.txt
certifi==2018.11.29
cffi==1.11.5
cloudpickle==0.6.1
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.0.0
decorator==4.3.0
imageio==2.4.1
kiwisolver==1.0.1
matplotlib==3.0.2
mkl-fft==1.0.6
mkl-random==1.0.2
networkx==2.2
numpy==1.15.4
olefile==0.46
opencv-python==3.4.5.20
Pillow==5.3.0
pycparser==2.19
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.12.0
toolz==0.9.0
torch==0.4.0
torchvision==0.2.0
tornado==5.1.1
tqdm==4.29.0
